#4 moved Display methods to a view package
