+++ src/com/company/controllers/TaskOperations.java
@@ -1,6 +1,8 @@
 package com.company.controllers;
 
 
+import com.company.view.Display;
+
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -37,7 +39,7 @@ public class TaskOperations {
     }
 
     public void editTask() throws IOException {
-        displayTaskList();
+        new Display().displayTaskList();
         System.out.print("Choose a Task to Edit > ");
         scanner = new Scanner(System.in);
         String fileName = scanner.nextLine() + ".txt";
@@ -113,35 +115,4 @@ public class TaskOperations {
 
 
     }
-
-
-    public void displayTaskList() {
-        //TODO implement sorting both by date and project
-        File currentDirectory = new File(".");
-
-        for (String fileName : currentDirectory.list()) {
-            if (fileName.endsWith(".txt")) {
-                System.out.println(fileName);
-            }
-        }
-    }
-
-    public void displayTask() {
-        displayTaskList();
-        scanner = new Scanner(System.in);
-        System.out.print("Enter project name to display >");
-        String projectName = scanner.next() + ".txt";
-
-        try (BufferedReader reader = new BufferedReader(new FileReader(projectName))) {
-            String line = null;
-            while ((line = reader.readLine()) != null) {
-                System.out.println(line);
-            }
-        } catch (FileNotFoundException e) {
-            System.err.println("File not found: " + projectName);
-        } catch (IOException e) {
-            System.out.println("Error reading file: " + projectName);
-        }
-    }
-
 }
